requirements.txt
certifi==2026.2.25
charset-normalizer==3.4.6
et_xmlfile==2.0.0
idna==3.11
openpyxl==3.1.5
python-dotenv==1.2.2
requests==2.32.5
urllib3==2.6.3
